Split schema file and flatten trees appropriately.
